refactored out functions that didnt belong in water_detector and added fast tiff reading function. Renamed sosi loader to generic io_tools
