+++ src/utils/io_tools.py
@@ -1,5 +1,9 @@
 from pathlib import Path
 import geopandas as gpd
+import numpy as np
+from osgeo import gdal
+import tifffile as tf
+
 from services.sosi_converter_service import convert_sosi_to_gpkg
 
 output_file = Path(__file__).parent.parent.parent /"tests" / "testdata" / "output.sos"
@@ -27,3 +31,21 @@ def get_gdf_content(gpkg_path: Path) -> gpd.GeoDataFrame:
     gdf = gpd.read_file(gpkg_path, layer="polygons", encoding="ISO-8859-1")
     return gdf
 
+def load_geotiff_dataset(path: str) ->  gdal.Dataset:
+    """
+    Load geotiff image into memory. Temporary function
+
+    :param path: path to the tiff image
+    :return: the gdal dataset.
+    """
+    ds = gdal.OpenEx(path)
+    return ds
+
+def read_tiff_fast(path) -> np.ndarray[tuple[int, int, int]]:
+    """
+    Fast reading of large tiff image using tifffile with turbojpeg. No metadata included, for that use
+    `load_geotiff_dataset`
+    :param path: path to the tiff image.
+    :return: the image as array in shape(bands, H, W).
+    """
+    return tf.imread(path, maxworkers=8)
